remove SeedFilter
