+++ RecoEgamma/EgammaElectronAlgos/src/ElectronSeedGenerator.cc
@@ -148,8 +148,8 @@ ElectronSeedGenerator::ElectronSeedGenerator(const edm::ParameterSet &pset, cons
       // new B/F configurables
       deltaPhi2B_(pset.getParameter<double>("DeltaPhi2B")),
       deltaPhi2F_(pset.getParameter<double>("DeltaPhi2F")),
-      phiMin2B_(pset.getParameter<double>("PhiMin2B")),
-      phiMin2F_(pset.getParameter<double>("PhiMin2F")),
+      phiMin2B_(-pset.getParameter<double>("PhiMax2B")),
+      phiMin2F_(-pset.getParameter<double>("PhiMax2F")),
       phiMax2B_(pset.getParameter<double>("PhiMax2B")),
       phiMax2F_(pset.getParameter<double>("PhiMax2F")),
       electronMatcher_(pset.getParameter<double>("ePhiMin1"),
@@ -158,11 +158,11 @@ ElectronSeedGenerator::ElectronSeedGenerator(const edm::ParameterSet &pset, cons
                        phiMax2B_,
                        phiMin2F_,
                        phiMax2F_,
-                       pset.getParameter<double>("z2MinB"),
+                       -pset.getParameter<double>("z2MaxB"),
                        pset.getParameter<double>("z2MaxB"),
-                       pset.getParameter<double>("r2MinF"),
+                       -pset.getParameter<double>("r2MaxF"),
                        pset.getParameter<double>("r2MaxF"),
-                       pset.getParameter<double>("rMinI"),
+                       -pset.getParameter<double>("rMaxI"),
                        pset.getParameter<double>("rMaxI"),
                        useRecoVertex_),
       positronMatcher_(pset.getParameter<double>("pPhiMin1"),
@@ -171,11 +171,11 @@ ElectronSeedGenerator::ElectronSeedGenerator(const edm::ParameterSet &pset, cons
                        phiMax2B_,
                        phiMin2F_,
                        phiMax2F_,
-                       pset.getParameter<double>("z2MinB"),
+                       -pset.getParameter<double>("z2MaxB"),
                        pset.getParameter<double>("z2MaxB"),
-                       pset.getParameter<double>("r2MinF"),
+                       -pset.getParameter<double>("r2MaxF"),
                        pset.getParameter<double>("r2MaxF"),
-                       pset.getParameter<double>("rMinI"),
+                       -pset.getParameter<double>("rMaxI"),
                        pset.getParameter<double>("rMaxI"),
                        useRecoVertex_) {}
