+++ RecoHI/HiEgammaAlgos/python/HiElectronSequence_cff.py
@@ -3,16 +3,15 @@ import FWCore.ParameterSet.Config as cms
 
 # creates the recoGsfTracks_electronGsfTracks__RECO = input GSF tracks
 from TrackingTools.GsfTracking.GsfElectronTracking_cff import *
-ecalDrivenElectronSeeds.SeedConfiguration.initialSeeds = "hiPixelTrackSeeds"
+ecalDrivenElectronSeeds.initialSeedsVector = cms.VInputTag(cms.InputTag("hiPixelTrackSeeds"))
 electronCkfTrackCandidates.src = "ecalDrivenElectronSeeds"
 
-ecalDrivenElectronSeeds.SeedConfiguration.maxHOverEBarrel = cms.double(0.25)
-ecalDrivenElectronSeeds.SeedConfiguration.maxHOverEEndcaps = cms.double(0.25)
+ecalDrivenElectronSeeds.maxHOverEBarrel = cms.double(0.25)
+ecalDrivenElectronSeeds.maxHOverEEndcaps = cms.double(0.25)
 
 electronGsfTrackingHiTask = cms.Task(ecalDrivenElectronSeeds ,
                                      electronCkfTrackCandidates ,
                                      electronGsfTracks)
-electronGsfTrackingHi = cms.Sequence(electronGsfTrackingHiTask)
 
 from RecoEgamma.EgammaElectronProducers.gsfElectronSequence_cff import *
