+++ RecoParticleFlow/PFTracking/plugins/modules.cc
@@ -4,7 +4,6 @@
 #include "RecoParticleFlow/PFTracking/interface/PFNuclearProducer.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFConversionProducer.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFV0Producer.h"
-#include "RecoParticleFlow/PFTracking/interface/ElectronSeedMerger.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFDisplacedVertexCandidateProducer.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFDisplacedVertexProducer.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFDisplacedTrackerVertexProducer.h"
@@ -17,7 +16,6 @@ DEFINE_FWK_MODULE(LightPFTrackProducer);
 DEFINE_FWK_MODULE(PFNuclearProducer);
 DEFINE_FWK_MODULE(PFConversionProducer);
 DEFINE_FWK_MODULE(PFV0Producer);
-DEFINE_FWK_MODULE(ElectronSeedMerger);
 DEFINE_FWK_MODULE(PFDisplacedVertexCandidateProducer);
 DEFINE_FWK_MODULE(PFDisplacedVertexProducer);
 DEFINE_FWK_MODULE(PFDisplacedTrackerVertexProducer);
